Add files via upload
This is an example of using K-Means Clustering to help identify seed type.  In this example I use the kmeans package imported from the sklearn library in python.  For this example I assume a known number of clusters (3) and specify that in the kmeans function.  I then fit the clusters to the data, assess the error and provide visualizations using principal component analysis to show misclassifications.